Fix warning "ordered comparison between pointer and integer"

+++ Source/arch/Intel.PC/32/include/Abstraction/Die.c
@@ -13,7 +13,7 @@ void DieFault(char* Message, Registers* Regs, bool Die)
 	(void)Die;
 	printf("Fault 0x%X, CS 0x%X, DS 0x%X, EAX 0x%X, EBP 0x%X, EBX 0x%X, ECX 0x%X, EDI 0x%X, EDX 0x%X, EFLAGS 0x%X, EIP 0x%X, INT 0x%X, ES 0x%X, ESI 0x%X, ESP 0x%X, FS 0x%X, GS 0x%X, SS 0x%X, USERESP 0x%X\r\n", Regs->Fault, Regs->CS, Regs->DS, Regs->EAX, Regs->EBP, Regs->EBX, Regs->ECX,
 	Regs->EDI, Regs->EDX, Regs->EFLAGS, Regs->EIP, Regs->InterruptNumber, Regs->ES, Regs->ESI, Regs->ESP, Regs->FS, Regs->GS, Regs->SS, Regs->UserESP);
-	if (Regs->EIP > &StartProgram && Regs->EIP < &EndProgram)
+	if (Regs->EIP > (unsigned int)&StartProgram && Regs->EIP < (unsigned int)&EndProgram)
 	{
 		printf("Program Range: Kernel <0x%X-0x%X>\r\n", &StartProgram, &EndProgram);
 	}
